CA-30853: Display proper error message when trying to eject from single host pool
It should say that this host is the master of the pool instead of
saying it is not a member of a pool.

Note: The only way you can tell the difference between a single host
pool and not being a pool member at the moment is to check
db.local_pool.name_label() which will be clear if not a pool member.