CA-41440: Replace 'Add/Remove DNS Servers' menu item with 'Display DNS Servers'
Remove the ability to modify the list of DNS servers as it was broken
and unneccessary.  The DNS configuration can be done from XenCenter or
from the CLI, and it is not neccessary for the DNS config to be correct
in order to get access via XenCenter or via ssh.

The reason this feature was broken is that it simply modified
/etc/resolv.conf, but this file is auto-generated.

For the record:
 /etc/resolv.conf is auto-generated by dhclient-script, or or by
 ifup-post  when it brings up the interface used for peer DNS (this is
 either the PIF with PIF.other-config.peerdns=true or the management
 PIF).  For static peer DNS interfaces the nameservers are taken from
 the comma seperated list in PIF.DNS.  For DHCP peer DNS interfaces
 the nameservers are those provided by the DHCP server, followed by
 the comma seperated list in PIF.DNS.