CA-43580: Improve xsconsole error message
The "Enter Maintenance Mode" feature can fail whilst migrating running
VMs if the other host(s) in the pool do not have enough memory
available.  In this circumstance XAPI will through a
"HOST_NOT_ENOUGH_FREE_MEMORY" error.

Previously this was translated by xsconsole as 'Not enough host memory
is available to perform this operation'.  This was confusing users,
because it didn't say what needed to be done to remedy the situation.

I have replaced this with the text 'Not enough host memory available,
please shutdown or suspend some VMs'.