[CP-18907] Eliminate spurious stack traces from log files
The `Data.Update` method was checking the validity of SR references
triggering a long list of hard-to-debug (but benign) errors during
xapi init phase. This can be often avoided by making sure that the
reference is not `OpaqueRef:NULL`. The additional log is to make
it easier to debug similar errors in the future.

I am assuming that we need to support Python 2.4, so I avoided
using the ternary operator.

Signed-off-by: Marcello Seri marcello.seri@citrix.com